Create the Virtusize interface
